All the capture replays
All of the test cases execute correctly fully except for the very last
one. I could not figure out how to make it select the tour type.  I did
some investigating and looked around at the commands, but I think the
original action is altered by code. I do not know where the code is
that is altering the html.